refactor(components folder): cleaned up all warnings after refactor of context
All components are now fully refactored to use context except for the onFinish functions.  I will
create a new branch to see about refactoring those later.  Everything is tested and working in the
same way it was before the refactor.  All warnings have been removed from the console at the time of
this commit.